Add EteSync to calendar and tasks sync
EteSync is a popular sync solution for contacts, calendars and tasks with clients for Android,
iOS, the desktop and the web. Including a DAV bridge to make it compatible with any DAV
complaint client. It's also featured on many other privacy lists.

As I said, it also supports contacts sync, though since there was no such category I didn't add
it, but should I?